Audio implemented and custom colors
The audio function has now been implemented, I added in configuration so
that the user can turn off the audio and adjust the volume. I also now
uploaded the config file and the graphs file. However the config file is
not the most up to date, so I will be changing it tomorrow. I got custom
colors working so the user can append "C:[R, G, B]" to the end of a
function and the graph will be colored accordingly. I also put the audio
playing on to a different thread. Now I just have to debug what dlls are
necessary for the program to boot.

+++ GraphingCalculator/include/Graph/Graph.h
@@ -8,9 +8,9 @@
 struct Point{
 	double x;
 	double y;
-	double r;
-	double g;
-	double b;
+	int r;
+	int g;
+	int b;
 	bool custom_color = false;
 	bool has_inequality = false;
 	bool inequality_result;
